Collapse repeated elisions (#42 ported to the ES6 codebase; closes #41) (#73)
* colorize: collapse multiple elisions into one line

Collapses >5 '...' into one line '... (6 entries)' (or similar)

* example: add big example to demonstrate elision

* test: add tests for collapsing elided objects

* colorize: push `options` down to `subcolorizeToCallback`

This means we can read the CLI/API-call options for the diffing from within
the colouriser, which is useful when we want to let the elision collapse
threshold be customizable by the user.

Also exposes `options` as an argument to `colorizeToArray`.

* Add 'maxElisions' option to customize collapse threshold

Instead of hardcoding the threshold to 5, let the user dynamically configure
it.  If not specified, repeated ...'s are never collapsed.

The new option is ignored without warning in 'full' mode.  This is intentional
so as to allow a user to 'alias jsondiff=jsondiff --max-elisions 5' or
similar.

Co-authored-by: Eric A Woudenberg <eaw@woudy.org>